backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
deepface==0.0.79
tensorflow==2.13.0
pymongo==4.6.0
python-dotenv==1.0.0